+++ jest.config.cjs
@@ -7,10 +7,11 @@ module.exports = {
     '^(\\.{1,2}/.*)\\.js$': '$1',
   },
   transform: {
-    '^.+\\.tsx?$': ['ts-jest', { useESM: true }],
+    '^.+\\.tsx?$': ['ts-jest', { useESM: true, tsconfig: { module: 'esnext' } }],
   },
   testMatch: ['<rootDir>/tests/**/*.test.ts'],
   setupFiles: ['<rootDir>/tests/helpers/envSetup.ts'],
   setupFilesAfterEnv: ['<rootDir>/tests/helpers/dbSetup.ts'],
-  testTimeout: 30000,
+  testTimeout: 60000,
+  forceExit: true,
 };